refactor(EG-941): de-duplicate run details step components across seqera and omics
