+++ packages/front-end/src/app/components/EGRunWorkflowFormReviewPipeline.vue
@@ -53,7 +53,7 @@
       const launchRequest: CreateWorkflowLaunchRequest = {
         launch: {
           computeEnvId: launchDetails.launch?.computeEnv?.id,
-          runName: wipSeqeraRun.value?.userPipelineRunName,
+          runName: wipSeqeraRun.value?.runName,
           pipeline: launchDetails.launch?.pipeline,
           revision: launchDetails.launch?.revision,
           configProfiles: launchDetails.launch?.configProfiles,
@@ -76,7 +76,7 @@
         const labRunRequest = {
           'LaboratoryId': wipSeqeraRun.value?.laboratoryId,
           'RunId': wipSeqeraRun.value?.transactionId,
-          'RunName': wipSeqeraRun.value?.userPipelineRunName,
+          'RunName': wipSeqeraRun.value?.runName,
           'Platform': 'Seqera Cloud', // TODO: Extend to support 'AWS HealthOmics',
           'Status': 'SUBMITTED',
           'WorkflowName': pipeline.value?.name, // TODO: Extend to support AWS HealthOmics Workflow name
@@ -148,7 +148,7 @@
         </div>
         <div class="text-md flex px-4 py-4">
           <dt class="w-48 text-black">Run Name</dt>
-          <dd class="text-muted text-left">{{ wipSeqeraRun?.userPipelineRunName }}</dd>
+          <dd class="text-muted text-left">{{ wipSeqeraRun?.runName }}</dd>
         </div>
       </dl>
     </section>
